Move more functions into view-utilities.ts.
