+++ src/plugins/dataflow/nodes/utilities/view-utilities.ts
@@ -20,3 +20,41 @@ export function getBoundingRectOfNodes(editor: NodeEditor): Rect | undefined {
   });
   return bounds;
 }
+
+export function getNewNodePosition(editor: NodeEditor) {
+  const kNodesPerColumn = 5;
+  const kNodesPerRow = 4;
+  const kColumnWidth = 200;
+  const kRowHeight = 90;
+  const kLeftMargin = 40;
+  const kTopMargin = 5;
+  const kColumnOffset = 15;
+
+  const numNodes = editor.nodes.length;
+  const { k } = editor.view.area.transform;
+  const nodePos: [number, number] =
+    [kLeftMargin * (1 / k) + Math.floor((numNodes % (kNodesPerColumn * kNodesPerRow)) / kNodesPerColumn)
+      * kColumnWidth + Math.floor(numNodes / (kNodesPerColumn * kNodesPerRow)) * kColumnOffset,
+    kTopMargin + numNodes % kNodesPerColumn * kRowHeight];
+  return nodePos;
+}
+
+export function moveNodeToFront(editor: NodeEditor, node: Node, newNode: boolean) {
+  const totalNodes = editor.nodes.length;
+  const selectedNodeView = editor.view.nodes.get(node);
+  const startingZIndex = selectedNodeView?.el.style.zIndex;
+  const selectedNodeZ = startingZIndex ? parseInt(startingZIndex, 10) : 0;
+  editor.nodes.forEach((n: Node) => {
+    const nodeView = editor.view.nodes.get(n);
+    if (nodeView) {
+      if (node.id === n.id) {
+        nodeView.el.style.zIndex = totalNodes.toString();
+      } else if (nodeView.el.style.zIndex) {
+        const nodeZ = parseInt(nodeView.el.style.zIndex, 10);
+        if (nodeZ > selectedNodeZ && !newNode) {
+          nodeView.el.style.zIndex = (nodeZ - 1).toString();
+        }
+      }
+    }
+  });
+}
